% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActProp_summary.R
\name{ActProp_summary}
\alias{ActProp_summary}
\title{Physical Activity Summary--Activity Proportions}
\usage{
ActProp_summary(final.dat)
}
\arguments{
\item{final.dat}{cleaned final data that is between record getup time and sleep time}
}
\value{
perc.sedentary Proportion of sedentary

perc.stand Proportion of standing

perc.step Proportion of stepping

step.per.day  Number of steps during the given period of time

break.per.day Number of interrupting sedentary behavior during the given period of time

break.rate Rate of interrupting sedentary behavior

MET.hour METs hours
}
\description{
Summarize sedantary, standing, and stepping activity measures
}
\details{
Proportion of sedentary/standing/stepping is the ratio of the total durations of sedentary/standing/step to the total worn time.

For number of steps, the number of stepping records multiplied by 2. This is because every two steps lead to one row's record in the event files.

Number of interrupting sedentary behavior is approximated by the number of  sedentary bouts.

Rate of interrupting sedentary behavior is the ratio of the daily number of interrupting sedentary behavior to total sedentary hours.

METs hours is the summation of METs hours from all activity records.
}
\examples{
 data(sampledata);ActProp_summary(sampledata)
}

