% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Simulated sample accelerometry data}
\format{A data frame with 129 rows and 7 variables}
\usage{
data(sample)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item Time. Time of this log row
  \item carat. weight of the diamond (0.2--5.01)
  \item cut. quality of the cut (Fair, Good, Very Good, Premium, Ideal)
}
}
\keyword{datasets}

