% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sed_summary.R
\name{sed_summary}
\alias{sed_summary}
\title{Physical Activity Summary--Sedentary Proportions and Percentiles}
\usage{
sed_summary(final.dat)
}
\arguments{
\item{final.dat}{cleaned final data that is between record getup time and sleep time}
}
\value{
total.sed.time Total sedantary hour, same as PA_summary function output sed.hour

total.number.of.sed.bouts Total number of sedentary bouts. It counts the number of sedentary recordings.

mean.sed.bout.length Mean sedentary bout length. It is the average of the duration time for all sedentary activities

prop.of.sed.time.greater.20min Proportions of sedentary bout greater than 20 minutes

prop.of.sed.time.greater.60min Proportions of sedentary bout greater than 60 minutes

prop.of.sed.time.greater.120min Proportions of sedentary bout greater than 120 minutes

total.sed.time.greater.20min Total sedentary time greater than 20 minutes

total.sed.time.greater.60min Total sedentary time greater than 60 minutes

total.sed.time.greater.120min Total sedentary time greater than 120 minutes

percentile.sed.time.5  5\% percentile of sedentary time

percentile.sed.time.25 25\% percentile of sedentary time

percentile.sed.time.50 50\% percentile of sedentary time

percentile.sed.time.75 75\% percentile of sedentary time

percentile.sed.time.95 95\% percentile of sedentary time

alpha.sed alpha of sendataty time, see details.

gini.index.sed  Gini index of sedantary, same as PA_summary function output gini.index

prop.sed.time.6.12 Proportions of sedentary time between 6:00-12:00

prop.sed.time.12.18 Proportions of sedentary time between 12:00-18:00

prop.sed.time.18.22 Proportions of sedentary time between 18:00-22:00
}
\description{
Summarize activity measures using proportions and percentiles
}
\details{
Proportions of sedentary bout greater than 20/60/120 minutes is the ratio of the number of sedentary bouts greater than 20/60/120 minutes to the total number of sedentary recordings.

Total sedentary time greater than 20/60/120 minutes is the summation of the sedentary durations which are greater than 20/60/120 minutes.

To calculate 5\%/25\%/.../95\% percentile of sedentary time, all of the recorded sedentary durations are listed and R function \emph{quantile} is used to find the percentiles.

alpha.sed is defined by \code{1+1/M}, where \code{M} is the average of \code{log(sedentary bout length /minimum sedentary bout length)}.

Proportions of sedentary time between 6:00-12:00/12:00-18:00/18:00-22:00 is the ratio of the sedentary durations to the total activity durations between 6:00-12:00/12:00-18:00/18:00-22:00.
}
\examples{
 data(sampledata);sed_summary(sampledata)
}

