% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLI_summary.R
\name{SLI_summary}
\alias{SLI_summary}
\title{Standing/light intensity Summary}
\usage{
SLI_summary(final_dat, takeoff.time, bed.time)
}
\arguments{
\item{final_dat}{cleaned final data that is between record getup time and sleep time}
}
\value{
\code{total_number_of_activity_bouts} Total events that are standing or stepping

\code{mean_activity_bout_length} Average length of activity bout

\code{prop_of_activity_time_greater_5min} The ratio of the number of active bouts greater than 5 min to the total number of active events

\code{prop_of_activity_time_greater_10min} The ratio of the number of active bouts greater than 10 min to the total number of active events.

\code{prop_of_activity_time_greater_30min} The ratio of the number of active bouts greater than 30 min to the total number of active events.

\code{total_activity_time_greater_5min} The summation of the active durations which are greater than 5 min

\code{total_activity_time_greater_10min} The summation of the active durations which are greater than 10 min

\code{total_activity_time_greater_30min} The summation of the active durations which are greater than 30 min

\code{percentile_activity_time_5} 5\% percentile of activity bouth length

\code{percentile_activity_time_25} 25\% percentile of activity bout length

\code{percentile_activity_time_50} 50\% percentile of activity bout length

\code{percentile_activity_time_75} 75\% percentile of activity bout length

\code{percentile_activity_time_95} 95\% percentile of activity bout length

\code{alpha_activity} alpha of activity time, see details

\code{stepping_to_standing_ratio} Ratio of stepping to standing, the ratio of total stepping hours to standing hours

\code{table} A table that combine all outputs
}
\description{
Summarize Standing and light intensity activity measures
}
\details{
\code{alpha_activity} is defined by \code{1+1/M}, where \code{M} is the average of \code{log(activity bout length /minimum activity bout length)}
}
\examples{
 data(sample_event);data(sample_bed_time);data(sample_takeon_time);SLI_summary(sample_event,sample_takeon_log,sample_bed_time)
}

