% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_summary.R
\name{Standard_summary}
\alias{Standard_summary}
\title{Standard Physical Activity Summary-Summations}
\usage{
Standard_summary(final_dat, takeoff.time, bed.time)
}
\arguments{
\item{final_dat}{cleaned final data that is between record getup time and sleep time}
}
\value{
\code{hours_worn_total} Total worn hours

\code{sed_hour}  Total sedantary hours

\code{stand_hour} Total stand hours

\code{step_hour} Total step hours

\code{step_count_total} Total Step count

\code{num_hour_over_3_METs} Number of hours that Metabolic Equivalent of Task (MET) is over 3

\code{MET_hours}  Total METs

\code{perc_sedentary} Proportion of sedentary

\code{perc_stand} Proportion of standing

\code{perc_step} Proportion of stepping

\code{step_per_day}  Number of steps during the given period of time

\code{table} A table that shows all above results
}
\description{
Summarize standard activity measures
}
\details{
All numbers are calculated in the given time period (day, hour, etc_)_ Total sedentary/standing/stepping hours are obtained from the summation of the duration times for sedentary/standing/stepping activities in the given time period_
}
\examples{
 data(sample_event);data(sample_bed_time);data(sample_takeon_time);Standard_summary(sample_event,sample_takeon_log,sample_bed_time)
}

