% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_consistent.R
\name{time_consistent}
\alias{time_consistent}
\title{Cut time outside range of log record}
\usage{
time_consistent(dat, record.start.time, record.end.time, ll)
}
\arguments{
\item{ll}{Row number of record get up time}
}
\value{
Data with right take on and take off time
}
\description{
Delete wrong data, all records should between log time for take on and take off. For the case that start in log time but last long outside the log, we just leave the time in the interval.
}

