% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{wear}
\alias{wear}
\title{Report participant's wear summary}
\usage{
wear(takeoff.time, bed.time, folder.to.export.APST.csvfile.and.RDatafile)
}
\arguments{
\item{takeoff.time}{The log file of taking off time}

\item{bed.time}{The log file of sleeping time}

\item{folder.to.export.APST.csvfile.and.RDatafile}{The folder to export files}

\item{n}{number of observations/participants}
}
\value{
the summary of wear
}
\description{
Import daily raw results and summarize them, export a summary table
}

