% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity_summary.R
\name{Activity_summary}
\alias{Activity_summary}
\title{Activity Summary--Standing and Stepping}
\usage{
Activity_summary(final.dat, record.getup.time, record.sleep.time, ll)
}
\arguments{
\item{final.dat}{cleaned final data that is between record getup time and sleep time}

\item{record.getup.time}{Log file of getup time}

\item{record.sleep.time}{Log file of sleep time}

\item{ll}{the id of getup time record}
}
\value{
... ...
}
\description{
Summarize Standing and Stepping activity measures
}

