% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA_summary.R
\name{PA_summary}
\alias{PA_summary}
\title{Physical Activity Summary}
\usage{
PA_summary(final.dat)
}
\arguments{
\item{final.dat}{cleaned final data that is between record getup time and sleep time}

\item{record.sleep.time}{Log file of sleep time}

\item{record.getup.time}{Log file of getup time}

\item{ll}{the id of getup time record}
}
\value{
hours.worn.total Total worn hours

hours.awake    Total awake hours

sed.hour  Total sedantary hours

... ...
}
\description{
Summarize activity measures
}
\examples{
 data(sample);PA_summary(sample) 
}

